Revert "Add boilerplate to develop new ImageSelect"
This reverts commit 6dc7613b484c9d6a116924e1a7e38562ee00fb1d.
