--- packages/components/src/image-select/index.js
@@ -1,2 +1,0 @@
-export { default as ImageSelect } from "./ImageSelect.js";
-export { default as ImageSelectButtons } from "./ImageSelectButtons.js";
